chore(ci) debug e2e test for quicker feedback loop (#1158)

+++ test/e2e/kuma_deploy_hybrid_kube_global_test.go
@@ -1,4 +1,4 @@
-package e2e
+package e2e_test
 
 import (
 	"strings"
@@ -65,6 +65,9 @@ var _ = Describe("Test Kubernetes/Universal deployment when Global is on K8S", f
 	})
 
 	AfterEach(func() {
+		if ShouldSkipCleanup() {
+			return
+		}
 		err := globalCluster.DeleteKuma(optsGlobal...)
 		Expect(err).ToNot(HaveOccurred())
 		err = globalCluster.DismissCluster()
